refac: create auth service
