+++ features/auth/services/authService.ts
@@ -1,0 +1,30 @@
+import apiInstance from "@/api/apiInstance";
+import { Credentials } from "@/features/auth/types/auth";
+import { useToast } from "@/hooks/useToast";
+import { useTokenStore } from "@/store/userData";
+import { useRouter } from "expo-router";
+import * as SecureStore from "expo-secure-store";
+
+const { setToken } = useTokenStore();
+const router = useRouter();
+const toast = useToast();
+
+export default async function authUser(credentials: Credentials) {
+  try {
+    const response = await apiInstance.post("/auth/login", credentials);
+    const token = response.data.token;
+
+    setToken(token);
+    await SecureStore.setItemAsync("user-token", token);
+    router.push("/(tabs)/");
+    toast({ heading: "Sucesso", message: "Seja bem-vindo!" });
+  } catch (e) {
+    console.error("User login failed: ", e);
+    //TODO: check error types to correctly handle them
+    toast({
+      heading: "Erro",
+      message: "Erro ao realizar login, por favor verifique suas credenciais.",
+      type: "error",
+    });
+  }
+}
