+++ features/auth/types/auth.ts
@@ -1,0 +1,4 @@
+export type Credentials = {
+  matricula: string;
+  senha: string;
+};
